requirements.txt
fastapi
uvicorn
asyncpg
scipy
yfinance
polygon-api-client
fpdf2
python-dotenv
alpaca-py
redis
